Solution: Create solution folder
