+++ pisek/task_jobs/solution/manager.py
@@ -15,27 +15,26 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import os
-import tempfile
-import time
 from typing import Any, Optional
 
 from pisek.jobs.jobs import State, Job, PipelineItemFailure
 from pisek.env.env import Env
 from pisek.utils.paths import TaskPath
-from pisek.config.config_types import TaskType, ProgramType, Scoring, DataFormat
+from pisek.config.config_types import TaskType, Scoring
 from pisek.utils.text import pad, pad_left, tab, POINTS_DEC_PLACES, format_points
 from pisek.utils.terminal import MSG_LEN, colored_env, right_aligned_text
-from pisek.task_jobs.verdicts_eval import evaluate_verdicts
+from pisek.task_jobs.solution.verdicts_eval import evaluate_verdicts
 from pisek.task_jobs.task_manager import TaskJobManager
-from pisek.task_jobs.program import RunResult, RunResultKind, ProgramsJob
 from pisek.task_jobs.compile import Compile
 from pisek.task_jobs.generator.input_info import InputInfo
-from pisek.task_jobs.data.data import OutputSmall
 from pisek.task_jobs.generator.manager import InputsInfoMixin
-from pisek.task_jobs.tools import IsClean
-from pisek.task_jobs.solution_result import Verdict, SolutionResult
+from pisek.task_jobs.solution.solution_result import Verdict, SolutionResult
 from pisek.task_jobs.judge import judge_job, RunJudge, RunCMSJudge, RunBatchJudge
+from pisek.task_jobs.solution.solution import (
+    RunSolution,
+    RunBatchSolution,
+    RunCommunication,
+)
 
 
 class SolutionManager(TaskJobManager, InputsInfoMixin):
@@ -121,7 +120,7 @@ class SolutionManager(TaskJobManager, InputsInfoMixin):
 
     def _create_batch_jobs(
         self, input_info: InputInfo, seed: int, subtask: int
-    ) -> tuple["RunBatchSolution", RunBatchJudge]:
+    ) -> tuple[RunBatchSolution, RunBatchJudge]:
         """Create RunSolution and RunBatchJudge jobs for batch task type."""
         input_path = input_info.task_path(self._env, seed)
         run_solution = RunBatchSolution(
@@ -146,7 +145,7 @@ class SolutionManager(TaskJobManager, InputsInfoMixin):
 
         return (run_solution, run_judge)
 
-    def _create_communication_jobs(self, inp: TaskPath) -> "RunCommunication":
+    def _create_communication_jobs(self, inp: TaskPath) -> RunCommunication:
         """Create RunCommunication job for communication task type."""
         if self._env.config.out_judge is None:
             raise RuntimeError("Unset judge for communication.")
@@ -455,142 +454,3 @@ class SubtaskJobGroup:
     def cancel(self):
         for job in self.new_run_jobs:
             job.cancel()
-
-
-RUN_JOB_NAME = r"Run (.*) on input (.*)"
-
-
-class RunSolution(ProgramsJob):
-    """Runs solution on given input."""
-
-    def __init__(
-        self,
-        env: Env,
-        name: str,
-        solution: TaskPath,
-        is_primary: bool,
-        **kwargs,
-    ) -> None:
-        super().__init__(env=env, name=name, **kwargs)
-        self.solution = solution
-        self.is_primary = is_primary
-
-    def _solution_type(self) -> ProgramType:
-        return (ProgramType.solve) if self.is_primary else (ProgramType.sec_solve)
-
-
-class RunBatchSolution(RunSolution):
-    def __init__(
-        self,
-        env: Env,
-        solution: TaskPath,
-        is_primary: bool,
-        input_: TaskPath,
-        output: Optional[TaskPath] = None,
-        **kwargs,
-    ) -> None:
-        name = RUN_JOB_NAME.replace(r"(.*)", solution.name, 1).replace(
-            r"(.*)", input_.name, 1
-        )
-        super().__init__(
-            env=env, name=name, solution=solution, is_primary=is_primary, **kwargs
-        )
-        self.input = input_
-        self.output = (
-            output
-            if output
-            else TaskPath.output_file(self._env, self.input.name, solution.name)
-        )
-        self.log_file = TaskPath.log_file(self._env, input_.name, self.solution.name)
-
-    def _run(self) -> RunResult:
-        return self._run_program(
-            program_type=self._solution_type(),
-            program=self.solution,
-            stdin=self.input,
-            stdout=self.output,
-            stderr=self.log_file,
-        )
-
-
-class RunCommunication(RunCMSJudge, RunSolution):
-    def __init__(
-        self,
-        env: Env,
-        solution: TaskPath,
-        is_primary: bool,
-        judge: TaskPath,
-        input_: TaskPath,
-        expected_points: Optional[float] = None,
-        **kwargs,
-    ):
-        super().__init__(
-            env=env,
-            name=RUN_JOB_NAME.replace(r"(.*)", solution.name, 1).replace(
-                r"(.*)", input_.name, 1
-            ),
-            judge=judge,
-            input_=input_,
-            expected_points=expected_points,
-            judge_log_file=TaskPath.log_file(
-                self._env, input_.name, f"{solution.name}.{judge.name}"
-            ),
-            solution=solution,
-            is_primary=is_primary,
-            **kwargs,
-        )
-        self.solution = solution
-        self.sol_log_file = TaskPath.log_file(
-            self._env, self.input.name, self.solution.name
-        )
-
-    def _get_solution_run_res(self) -> RunResult:
-        with tempfile.TemporaryDirectory() as fifo_dir:
-            fifo_from_solution = os.path.join(fifo_dir, "solution-to-manager")
-            fifo_to_solution = os.path.join(fifo_dir, "manager-to-solution")
-
-            os.mkfifo(fifo_from_solution)
-            os.mkfifo(fifo_to_solution)
-
-            pipes = [
-                os.open(fifo_from_solution, os.O_RDWR),
-                os.open(fifo_to_solution, os.O_RDWR),
-                # Open fifos to prevent blocking on future opens
-                fd_from_solution := os.open(fifo_from_solution, os.O_WRONLY),
-                fd_to_solution := os.open(fifo_to_solution, os.O_RDONLY),
-            ]
-
-            self._load_program(
-                ProgramType.judge,
-                self.judge,
-                stdin=self.input,
-                stdout=self.points_file,
-                stderr=self.judge_log_file,
-                args=[fifo_from_solution, fifo_to_solution],
-            )
-
-            self._load_program(
-                self._solution_type(),
-                self.solution,
-                stdin=fd_to_solution,
-                stdout=fd_from_solution,
-                stderr=self.sol_log_file,
-            )
-
-            def close_pipes(_):
-                time.sleep(0.05)
-                for pipe in pipes:
-                    os.close(pipe)
-
-            self._load_callback(close_pipes)
-
-            judge_res, sol_res = self._run_programs()
-            self._judge_run_result = judge_res
-
-            return sol_res
-
-    def _judge(self) -> SolutionResult:
-        return self._load_solution_result(self._judge_run_result)
-
-    def _judging_message(self) -> str:
-        return f"solution {self.solution:p} on input {self.input:p}"
