+++ pisek/task_jobs/solution/solution_result.py
