+++ pisek/task_jobs/testing_log.py
@@ -22,7 +22,7 @@ from pisek.task_jobs.task_manager import (
     TaskJobManager,
     SOLUTION_MAN_CODE,
 )
-from pisek.task_jobs.solution_result import SolutionResult
+from pisek.task_jobs.solution.solution_result import SolutionResult
 
 TESTING_LOG = "testing_log.json"
